+++ app/forms/decidim/decidim_awesome/admin/config_form.rb
@@ -11,8 +11,6 @@ module Decidim
         attribute :allow_images_in_editors, Boolean
         attribute :allow_videos_in_editors, Boolean
         attribute :allow_images_in_proposals, Boolean
-        attribute :use_markdown_editor, Boolean
-        attribute :allow_images_in_markdown_editor, Boolean
         attribute :auto_save_forms, Boolean
         attribute :scoped_styles, Hash
         attribute :proposal_custom_fields, Hash