+++ app/packs/src/decidim/decidim_awesome/awesome_admin.js
@@ -10,6 +10,4 @@ import "src/decidim/decidim_awesome/admin/form_exit_warn"
 import "src/decidim/decidim_awesome/proposals/custom_fields"
 import "src/decidim/decidim_awesome/admin/custom_fields_builder"
 
-// import "src/decidim/decidim_awesome/editors/tabs_focus"
-
 window.DecidimAwesome = window.DecidimAwesome || {};