+++ lib/decidim/decidim_awesome/test/initializer.rb
@@ -6,8 +6,6 @@ Decidim::DecidimAwesome.configure do |config|
       :allow_images_in_editors,
       :allow_videos_in_editors,
       :allow_images_in_proposals,
-      :use_markdown_editor,
-      :allow_images_in_markdown_editor,
       :auto_save_forms,
       :intergram_for_admins,
       :intergram_for_public,