+++ lib/decidim/decidim_awesome/test/shared_examples/editor_examples.rb
@@ -53,37 +53,3 @@ shared_examples "has drag and drop" do
     expect(proposal.reload.body["en"]).to include(last_image.attached_uploader(:file).path)
   end
 end
-
-shared_examples "has markdown editor" do |images|
-  it "has CodeMirror class" do
-    skip "This feature is pending to be adapted to Decidim 0.28"
-
-    expect(page).to have_xpath("//div[@class='CodeMirror cm-s-paper CodeMirror-wrap']")
-  end
-
-  it "has toolbar" do
-    expect(page).to have_xpath("//div[@class='editor-toolbar']")
-  end
-
-  if images
-    it "has help text" do
-      skip "This feature is pending to be adapted to Decidim 0.28"
-
-      expect(page).to have_content("Add images by dragging & dropping or pasting them.")
-    end
-  else
-    it "has no help text" do
-      expect(page).to have_no_content("Add images by dragging & dropping or pasting them.")
-    end
-  end
-end
-
-shared_examples "has no markdown editor" do
-  it "has CodeMirror class" do
-    expect(page).to have_no_xpath("//div[@class='CodeMirror cm-s-paper CodeMirror-wrap']")
-  end
-
-  it "has toolbar" do
-    expect(page).to have_no_xpath("//div[@class='editor-toolbar']")
-  end
-end