+++ spec/system/public/edit_proposals_spec.rb
@@ -20,13 +20,9 @@ describe "Show proposals editor" do
   let!(:allow_images_in_proposals) { create(:awesome_config, organization:, var: :allow_images_in_proposals, value: images_in_proposals) }
   let!(:allow_images_in_editors) { create(:awesome_config, organization:, var: :allow_images_in_editors, value: images_editor) }
   let!(:allow_videos_in_editors) { create(:awesome_config, organization:, var: :allow_videos_in_editors, value: videos_editor) }
-  let!(:use_markdown_editor) { create(:awesome_config, organization:, var: :use_markdown_editor, value: markdown_enabled) }
-  let!(:allow_images_in_markdown_editor) { create(:awesome_config, organization:, var: :allow_images_in_markdown_editor, value: markdown_images) }
   let(:images_in_proposals) { false }
   let(:images_editor) { false }
   let(:videos_editor) { false }
-  let(:markdown_enabled) { false }
-  let(:markdown_images) { false }
   let(:editor_selector) { "textarea#proposal_body" }
   let(:image) { Decidim::Dev.test_file("city.jpeg", "image/jpeg") }
 
@@ -64,18 +60,6 @@ describe "Show proposals editor" do
       it_behaves_like "has image support"
       it_behaves_like "has video support"
     end
-
-    # context "and markdown is enabled" do
-    #   let(:markdown_enabled) { true }
-
-    #   it_behaves_like "has markdown editor"
-
-    #   context "and images are enabled" do
-    #     let(:markdown_images) { true }
-
-    #     it_behaves_like "has markdown editor", true
-    #   end
-    # end
   end
 
   context "when rich text editor is NOT enabled for participants" do
@@ -86,30 +70,5 @@ describe "Show proposals editor" do
 
       it_behaves_like "has drag and drop"
     end
-
-    # context "and markdown is enabled" do
-    #   let(:markdown_enabled) { true }
-
-    #   it_behaves_like "has no markdown editor"
-    # end
   end
-
-  # context "when editing in markdown mode" do
-  #   let(:rte_enabled) { true }
-  #   let(:markdown_enabled) { true }
-  #   let(:text) { "# title\\n\\nParagraph\\nline 2" }
-  #   let(:html) { "<h1 id=\"title\">title</h1><p>Paragraph<br>line 2</p>" }
-
-  #   it "converts markdown to html before saving" do
-  #     sleep 1
-  #     page.execute_script("$('[name=\"faker-inscrybmde\"]:first')[0].InscrybMDE.value('#{text}')")
-
-  #     click_link_or_button "Send"
-
-  #     expect(page).to have_no_content("# title")
-  #     expect(page).to have_css("h1", text: "title")
-  #     expect(page).to have_css("p", text: "Paragraph\nline 2")
-  #     expect(Decidim::Proposals::Proposal.last.body["en"].gsub(/[\n\r]/, "")).to eq(html)
-  #   end
-  # end
 end